fix: strip .git from repo URL before appending
It's legal for a repository URL to end with .git, but cargo-dist currently assumes
it's a web URL rather than a git clone URL before working with it. This makes sure we
strip the .git part of the URL before tacking on the `/release` path so we
generate a valid URL.

Fixes #263.